move tests, add more complex tests, comment where broke
