+++ packages/xlucene-parser/test/cases/range-relative-dates.ts
@@ -1,28 +1,10 @@
-import { xLuceneFieldType, xLuceneTypeConfig, xLuceneVariables } from '@terascope/types';
+import { xLuceneFieldType } from '@terascope/types';
 import subDays from 'date-fns/subDays';
 import addDays from 'date-fns/addDays';
-import {
-    NodeType, Node
-    //  Range, RangeNode
-} from '../../src';
-// import { TestCase } from './interfaces';
+import { NodeType, Node } from '../../src';
+import { TestCase } from './interfaces';
 
-type TestCase = [
-    // when give query %s
-    string,
-    // it should be able to parse %s
-    string,
-    // toMatchObject(%j)
-    Partial<Node>|string,
-    // Type config to pass in
-    xLuceneTypeConfig?,
-    xLuceneVariables?,
-    // to test resolveVariables separately
-    Partial<Node>?,
-    ((val: Date, ast: Node) => any)?
-];
-
-export default [
+export const relativeDateRanges = [
     [
         'val:[-3d TO 2d]',
         'should NOT coerce to relative date to date when no type config provided',
@@ -75,7 +57,7 @@ export default [
                 field_type: xLuceneFieldType.Date,
                 value: { type: 'value' } // value tested below
             }
-        },
+        } as Node,
         { val: xLuceneFieldType.Date },
         undefined, undefined,
         (now: Date, ast: any) => {
@@ -143,34 +125,6 @@ export default [
             expect(astEndDate).toEqual(addDays(_now, 21));
         }
     ],
-    // [
-    //     'val:[2012-01-01||+1M/d TO now+20d-3d-1d+5d]',
-    //     'date math range anchored',
-    //     {
-    //         type: NodeType.Range,
-    //         field: 'val',
-    //         left: {
-    //             operator: 'gte',
-    //             field_type: xLuceneFieldType.Date,
-    //             value: { type: 'value' } // value tested below
-    //         },
-    //         right: {
-    //             operator: 'lte',
-    //             field_type: xLuceneFieldType.Date,
-    //             value: { type: 'value' } // value tested below
-    //         }
-    //     },
-    //     { val: xLuceneFieldType.Date },
-    //     undefined, undefined,
-    //     (now: Date, ast: any) => {
-    //         const _now = new Date(now.setUTCHours(0, 0, 0, 0));
-    //         const astStartDate = new Date(ast.left.value.value.setUTCHours(0, 0, 0, 0));
-    //         const astEndDate = new Date(ast.right.value.value.setUTCHours(0, 0, 0, 0));
-
-    //         expect(astStartDate).toEqual(addDays(_now, 6));
-    //         expect(astEndDate).toEqual(addDays(_now, 21));
-    //     }
-    // ],
     [
         'val:[now TO now+5D]',
         'date math range with left now',
@@ -367,15 +321,155 @@ export default [
             expect(astEndDate).toEqual(subDays(_now, 2));
         }
     ],
+    [
+        'val:[$start TO $end]',
+        'variables - relative and date math range mixed capitalization and extra spaces',
+        {
+            type: NodeType.Range,
+            field: 'val',
+            left: {
+                operator: 'gte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            },
+            right: {
+                operator: 'lte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            }
+        },
+        { val: xLuceneFieldType.Date },
+        { start: '-3   dAyS', end: 'nO W+  1d ' },
+        undefined,
+        (now: Date, ast: any) => {
+            const _now = new Date(now.setUTCHours(0, 0, 0, 0));
+            const astStartDate = new Date(ast.left.value.value.setUTCHours(0, 0, 0, 0));
+            const astEndDate = new Date(ast.right.value.value.setUTCHours(0, 0, 0, 0));
+
+            expect(astStartDate).toEqual(subDays(_now, 3));
+            expect(astEndDate).toEqual(addDays(_now, 1));
+        }
+    ],
+    [
+        'val:[$start TO now+2d]',
+        'variable only for left range',
+        {
+            type: NodeType.Range,
+            field: 'val',
+            left: {
+                operator: 'gte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            },
+            right: {
+                operator: 'lte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            }
+        },
+        { val: xLuceneFieldType.Date },
+        { start: '-3   dAyS', end: 'nO W+  1d ' },
+        undefined,
+        (now: Date, ast: any) => {
+            const _now = new Date(now.setUTCHours(0, 0, 0, 0));
+            const astStartDate = new Date(ast.left.value.value.setUTCHours(0, 0, 0, 0));
+            const astEndDate = new Date(ast.right.value.value.setUTCHours(0, 0, 0, 0));
+
+            expect(astStartDate).toEqual(subDays(_now, 3));
+            expect(astEndDate).toEqual(addDays(_now, 2));
+        }
+    ],
+    [
+        'val:[2021-03-01 TO $end]',
+        'variable only for right range',
+        {
+            type: NodeType.Range,
+            field: 'val',
+            left: {
+                operator: 'gte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            },
+            right: {
+                operator: 'lte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            }
+        },
+        { val: xLuceneFieldType.Date },
+        { start: '-3   dAyS', end: 'nO W+  3d ' },
+        undefined,
+        (now: Date, ast: any) => {
+            const _now = new Date(now.setUTCHours(0, 0, 0, 0));
+            const astStartDate = ast.left.value.value;
+            const astEndDate = new Date(ast.right.value.value.setUTCHours(0, 0, 0, 0));
+
+            expect(astStartDate).toEqual('2021-03-01');
+            expect(astEndDate).toEqual(addDays(_now, 3));
+        }
+    ],
+    [
+        'val:[$start TO $end]',
+        'should not convert if no type config is provided',
+        {
+            type: NodeType.Range,
+            field: 'val',
+            left: {
+                operator: 'gte',
+                type: NodeType.Term,
+                value: { type: 'variable', scoped: false, value: 'start' }
+            },
+            right: {
+                operator: 'lte',
+                type: NodeType.Term,
+                value: { type: 'variable', scoped: false, value: 'end' }
+            }
+        } as Node,
+        undefined,
+        { start: '-3   days', end: 'now+1d ' },
+    ],
+    [
+        'val:[$start TO $end]',
+        'variable more complex date math - division and anchor add',
+        {
+            type: NodeType.Range,
+            field: 'val',
+            left: {
+                operator: 'gte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            },
+            right: {
+                operator: 'lte',
+                field_type: xLuceneFieldType.Date,
+                value: { type: 'value' } // value tested below
+            }
+        },
+        { val: xLuceneFieldType.Date },
+        { start: 'now-4d/y', end: '2021-01-02||+4d' },
+        undefined,
+        (now: Date, ast: any) => {
+            // /y should make it start of year
+            expect(new Date(ast.left.value.value).toISOString()).toInclude('01-01T00:00:00.000');
+            // should be 1-2-21 plus 4 days
+            expect(new Date(ast.right.value.value).toISOString()).toInclude('2021-01-06T00:00:00.000');
+        }
+    ],
 ] as TestCase[];
 
-export const failures = [
+export const relativeDateRangeFailures = [
+    [
+        'val:[2021-03-01 TO $end]',
+        'when variable not provided',
+        'Expected undefined (undefined) to be in a date like format',
+        { val: xLuceneFieldType.Date },
+        { srt: '-3   dAyS', ed: 'nO W+  3d ' },
+    ],
     [
         'val:[nww TO -3d]',
         'when date math "now" is misspelled',
         'Failure to parse xLucene query "val:[nww TO -3d]", caused by Error: Expected nww (String) to be in a date like format',
         { val: xLuceneFieldType.Date },
-
     ],
     [
         'val:[now TO 3dys]',
@@ -395,19 +489,4 @@ export const failures = [
         'Failure to parse xLucene query "val:[now TO 3 days]", caused by SyntaxError: Expected "]", "}", or a character between 0-9 but " " found.',
         { val: xLuceneFieldType.Date },
     ],
-] as TestCase[];
-
-// function testDates(now: Date, ast: any, start: any, end: any) {
-//     const _now = new Date(now.setUTCHours(0, 0, 0, 0));
-//     const astStartDate = new Date(ast.left.value.value.setUTCHours(0, 0, 0, 0));
-//     const astEndDate = new Date(ast.right.value.value.setUTCHours(0, 0, 0, 0));
-
-//     expect(astStartDate).toEqual(subDays(_now, 3));
-//     expect(astEndDate).toEqual(addDays(_now, 2));
-// }
-
-/**
- * TODO -
- * - add tests w/variables
- * - maybe replacing white space so last failure is fine
- */
+];
